requirements.txt
streamlit
yfinance
pandas
numpy
requests
beautifulsoup4
snownlp
scikit-learn
plotly
tensorflow-cpu